backend/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Django==1.11.5
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.1
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gensim
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
PyJWT==1.5.3
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
six==1.11.0
sqlparse==0.2.3
urllib3==1.22
whitenoise==3.3.1
